+++ LiveSegmentation/LiveSegmentation.py
@@ -13,7 +13,7 @@ from slicer.ScriptedLoadableModule import (
     ScriptedLoadableModuleWidget,
 )
 
-PLUGIN_VERSION = "0.11.2"
+PLUGIN_VERSION = "0.11.3"
 
 
 class LiveSegmentation(ScriptedLoadableModule):
@@ -718,4 +718,4 @@ class LiveSegmentationWidget(ScriptedLoadableModuleWidget):
 class LiveSegmentationTest(ScriptedLoadableModuleTest):
     def runTest(self):
         self.delayDisplay("Live Segmentation module loaded")
-        self.assertEqual(PLUGIN_VERSION, "0.11.2")
+        self.assertEqual(PLUGIN_VERSION, "0.11.3")